+++ com.chartboost.mediation/Editor/Adapters/Serialization/Documentation.cs
@@ -2,11 +2,20 @@ using System;
 
 namespace Chartboost.Editor.Adapters.Serialization
 {
+    /// <summary>
+    /// Contains references to Adapters Documentation.
+    /// </summary>
     [Serializable]
     public struct Documentation
     {
+        /// <summary>
+        /// Chartboost specific documentation.
+        /// </summary>
         public string chartboost;
-
+        
+        /// <summary>
+        /// Partner network specific documentation.
+        /// </summary>
         public string partner;
     }
 }