+++ com.chartboost.mediation/Editor/Adapters/Serialization/NativeAdapter.cs
@@ -2,6 +2,9 @@ using System;
 
 namespace Chartboost.Editor.Adapters.Serialization
 {
+    /// <summary>
+    /// Base structure for native adapters. 
+    /// </summary>
     #nullable enable
     [Serializable]
     public struct NativeAdapter
@@ -12,29 +15,14 @@ namespace Chartboost.Editor.Adapters.Serialization
         public string adapter;
 
         /// <summary>
-        /// SDK dependency name of the Native Ad Adapter.
-        /// </summary>
-        public string sdk;
-        
-        /// <summary>
-        /// Additional dependencies for the certified Ad Adapter
-        /// </summary>
-        public string[]? dependencies;
-        
-        /// <summary>
         /// Additional repositories to fetch dependencies. Some providers have their own repos, they can be found here.
         /// </summary>
         public string[]? repositories;
-
+        
         /// <summary>
         /// All certified versions of the Ad Adapter
         /// </summary>
-        public string[] versions;
-
-        /// <summary>
-        /// Indicates if Native SDK dependency should be added to all targets, this is defined on a case by case basis.
-        /// </summary>
-        public bool? allTargets;
+        public AdapterVersion[]? versions;
     }
     #nullable disable
 }