+++ com.chartboost.mediation/Editor/Adapters/Serialization/Platform.cs
@@ -1,11 +1,15 @@
+using System;
+
 namespace Chartboost.Editor.Adapters.Serialization
 {
     /// <summary>
-    /// Supported platforms for Adapters.
+    /// Supported platforms for Adapters. Flag enabled.
     /// </summary>
+    [Flags]
     public enum Platform
     {
-        Android,
-        IOS
+        None = 0,
+        Android = 1,
+        IOS = 2
     }
 }