+++ com.chartboost.mediation/Editor/Adapters/Serialization/SDKSelections.cs
@@ -2,12 +2,21 @@ using System;
 
 namespace Chartboost.Editor.Adapters.Serialization
 {
+    /// <summary>
+    /// Root structure for User selections. Saved into a json file.
+    /// </summary>
     #nullable enable
     [Serializable]
     public class SDKSelections
     {
+        /// <summary>
+        /// Current Chartboost Mediation version dependencies, this value is generated off the Chartboost Mediation UPM package.
+        /// </summary>
         public string? mediationVersion;
 
+        /// <summary>
+        /// User specific Ad Adapter selections.
+        /// </summary>
         public AdapterSelection[]? adapterSelections;
     }
     #nullable disable