+++ com.chartboost.mediation/Runtime/FullScreen/ChartboostMediationFullScreenBase.cs
@@ -5,49 +5,49 @@ namespace Chartboost.FullScreen
     /// </summary>
     public abstract class ChartboostMediationFullScreenBase : IChartboostMediationAd, IChartboostMediationFullScreenAd
     {
-        protected static string LogTag = "ChartboostMediationFullScreen (Base)";
-        protected readonly string PlacementName;
+        protected static string logTag = "ChartboostMediationFullScreen (Base)";
+        protected readonly string placementName;
 
         protected ChartboostMediationFullScreenBase(string placementName)
-            => PlacementName = placementName;
+            => this.placementName = placementName;
 
         /// <inheritdoc cref="IChartboostMediationAd.SetKeyword"/>>
         public virtual bool SetKeyword(string keyword, string value)
         {
-            Logger.Log(LogTag, $"fullscreen: {PlacementName}, setting keyword {keyword} with value: {value}");
+            Logger.Log(logTag, $"fullscreen: {placementName}, setting keyword {keyword} with value: {value}");
             return false;
         }
 
         /// <inheritdoc cref="IChartboostMediationAd.RemoveKeyword"/>>
         public virtual string RemoveKeyword(string keyword)
         {
-            Logger.Log(LogTag, $"fullscreen: {PlacementName}, removing keyword {keyword}");
+            Logger.Log(logTag, $"fullscreen: {placementName}, removing keyword {keyword}");
             return keyword;
         }
 
         /// <inheritdoc cref="IChartboostMediationAd.Destroy"/>>
         public virtual void Destroy()
-            => Logger.Log(LogTag, $"destroying fullscreen: {PlacementName}");
+            => Logger.Log(logTag, $"destroying fullscreen: {placementName}");
 
         /// <inheritdoc cref="IChartboostMediationFullScreenAd.Load"/>>
         public virtual void Load()
-            => Logger.Log(LogTag, $"loading fullscreen: {PlacementName}");
+            => Logger.Log(logTag, $"loading fullscreen: {placementName}");
 
         /// <inheritdoc cref="IChartboostMediationFullScreenAd.Show"/>>
         public virtual void Show()
-            => Logger.Log(LogTag, $"showing fullscreen: {PlacementName}");
+            => Logger.Log(logTag, $"showing fullscreen: {placementName}");
 
         /// <inheritdoc cref="IChartboostMediationFullScreenAd.ReadyToShow"/>>
         public virtual bool ReadyToShow()
         {
-            Logger.Log(LogTag, $"checking fullscreen: {PlacementName} availability");
+            Logger.Log(logTag, $"checking fullscreen: {placementName} availability");
             return false;
         }
 
         /// <inheritdoc cref="IChartboostMediationFullScreenAd.ClearLoaded"/>>
         public virtual void ClearLoaded()
         {
-            Logger.Log(LogTag, $"clearing fullscreen: {PlacementName}");
+            Logger.Log(logTag, $"clearing fullscreen: {placementName}");
         }
     }
     
@@ -58,7 +58,7 @@ namespace Chartboost.FullScreen
     {
         public ChartboostMediationInterstitialUnsupported(string placementName) : base(placementName)
         {
-            LogTag = "ChartboostMediationInterstitial (Unsupported)";
+            logTag = "ChartboostMediationInterstitial (Unsupported)";
         }
     }
 }