+++ com.chartboost.mediation/Runtime/Plugins/iOS/ChartboostMediationBinding.m
@@ -26,10 +26,6 @@ static void sendToMain(block block) {
     dispatch_async(dispatch_get_main_queue(), block);
 }
 
-static void sendToBackground(block block) {
-    dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0), block);
-}
-
 void _setLifeCycleCallbacks(ChartboostMediationEvent didStartCallback, ChartboostMediationILRDEvent didReceiveILRDCallback, ChartboostMediationPartnerInitializationDataEvent didReceivePartnerInitializationDataCallback)
 {
     [[ChartboostMediationManager sharedManager] setLifeCycleCallbacks:didStartCallback didReceiveILRDCallback:didReceiveILRDCallback didReceivePartnerInitializationData:didReceivePartnerInitializationDataCallback];
@@ -116,7 +112,7 @@ char * _chartboostMediationInterstitialRemoveKeyword(const void *uniqueId, const
 
 void _chartboostMediationInterstitialAdLoad(const void * uniqueId)
 {
-    sendToBackground(^{
+    sendToMain(^{
         id<HeliumInterstitialAd> ad = (__bridge id<HeliumInterstitialAd>)uniqueId;
         [ad loadAd];
     });
@@ -130,7 +126,7 @@ void _chartboostMediationInterstitialClearLoaded(const void * uniqueId)
 
 void _chartboostMediationInterstitialAdShow(const void * uniqueId)
 {
-    sendToBackground(^{
+    sendToMain(^{
         id<HeliumInterstitialAd> ad = (__bridge id<HeliumInterstitialAd>)uniqueId;
         [ad showAdWithViewController: UnityGetGLViewController()];
     });
@@ -140,13 +136,6 @@ BOOL _chartboostMediationInterstitialAdReadyToShow(const void * uniqueId)
 {
     id<HeliumInterstitialAd> ad = (__bridge id<HeliumInterstitialAd>)uniqueId;
     return [ad readyToShow];
-}
-
-void _chartboostMediationFreeInterstitialAdObject(const void * uniqueId)
-{
-    sendToBackground(^{
-        [[ChartboostMediationManager sharedManager] freeInterstitialAd: [NSNumber numberWithLong:(long)uniqueId]];
-    });
 }
 
 void * _chartboostMediationGetRewardedAd(const char *placementName)
@@ -170,7 +159,7 @@ char * _chartboostMediationRewardedRemoveKeyword(const void *uniqueId, const cha
 
 void _chartboostMediationRewardedAdLoad(const void * uniqueId)
 {
-    sendToBackground(^{
+    sendToMain(^{
         id<HeliumRewardedAd> ad = (__bridge id<HeliumRewardedAd>)uniqueId;
         [ad loadAd];
     });
@@ -184,7 +173,7 @@ void _chartboostMediationRewardedClearLoaded(const void * uniqueId)
 
 void _chartboostMediationRewardedAdShow(const void * uniqueId)
 {
-    sendToBackground(^{
+    sendToMain(^{
         id<HeliumRewardedAd> ad = (__bridge id<HeliumRewardedAd>)uniqueId;
         [ad showAdWithViewController: UnityGetGLViewController()];
     });
@@ -200,13 +189,6 @@ void _chartboostMediationRewardedAdSetCustomData(const void * uniqueId, const ch
 {
     id<HeliumRewardedAd> ad = (__bridge id<HeliumRewardedAd>)uniqueId;
     ad.customData = GetStringParam(customData);
-}
-
-void _chartboostMediationFreeRewardedAdObject(const void * uniqueId)
-{
-    sendToBackground(^{
-        [[ChartboostMediationManager sharedManager] freeRewardedAd: [NSNumber numberWithLong:(long)uniqueId]];
-    });
 }
 
 void * _chartboostMediationGetBannerAd(const char *placementName, long size)
@@ -342,16 +324,18 @@ void _chartboostMediationBannerDisableDrag(const void * uniqueId){
 
 void _chartboostMediationBannerClearLoaded(const void * uniqueId)
 {
-    id<HeliumBannerAd> ad = (__bridge id<HeliumBannerAd>)uniqueId;
-    [ad clearAd];
+    sendToMain(^(){
+        id<HeliumBannerAd> ad = (__bridge id<HeliumBannerAd>)uniqueId;
+        [ad clearAd];
+    });
 }
 
 void _chartboostMediationBannerRemove(const void * uniqueId)
 {
-    sendToMain(^{
+    sendToMain(^(){
         HeliumBannerView* bannerView = (__bridge HeliumBannerView*)uniqueId;
         [bannerView removeFromSuperview];
-        [[ChartboostMediationManager sharedManager] freeBannerAd: [NSNumber numberWithLong:(long)uniqueId]];
+        [[ChartboostMediationManager sharedManager] freeAd: [NSNumber numberWithLong:(long)uniqueId] placementName:nil multiPlacementSupport:true];
     });
 }
 
@@ -363,10 +347,9 @@ void _chartboostMediationBannerSetVisibility(const void * uniqueId, BOOL isVisib
     });
 }
 
-void _chartboostMediationFreeBannerAdObject(const void * uniqueId)
+void _chartboostMediationFreeAdObject(const void * uniqueId, const char * placementName, bool multiPlacementSupport)
 {
-    sendToBackground(^{
-        [[ChartboostMediationManager sharedManager] freeBannerAd: [NSNumber numberWithLong:(long)uniqueId]];
+    sendToMain(^(){
+        [[ChartboostMediationManager sharedManager] freeAd: [NSNumber numberWithLong:(long)uniqueId] placementName:GetStringParam(placementName) multiPlacementSupport:multiPlacementSupport];
     });
-}
- 
+}